engine/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
openai==1.76.*
pyyaml==6.0.*
python-dotenv==1.1.*
slowapi==0.1.*
httpx==0.28.*
